[#183] util unittests
