+++ src/ml_service/init_test.py
@@ -1,9 +1,8 @@
 import os
 import unittest
-
-from util.constant import Global
-
+from util.constant import Path
 
 class TestStringMethods(unittest.TestCase):
     def test_init(self):
-        self.assertTrue(os.path.isfile(Global.word_vector_directory + 'non-lem.bin'))
+        import init
+        self.assertTrue(os.path.isfile(Path.binary_directory + 'non-lem.bin'))
