+++ src/ml_service/util/log.py
@@ -1,6 +1,5 @@
 import logging
 import os
-
 from util import constant
